requirements.txt
Pillow==6.1.0
pymongo==3.8.0
mongoengine==0.10.7
pandas==0.20.3